restructure: tier F — group bytecode compiler under compile/
