+++ jetro-core/src/lib.rs
@@ -38,8 +38,7 @@ pub(crate) mod planner;
 pub(crate) mod structural;
 pub(crate) mod util;
 pub(crate) mod view_pipeline;
-pub(crate) mod compiler;
-pub(crate) mod compiler_passes;
+pub(crate) mod compile;
 pub(crate) mod vm;
 pub(crate) mod logical_plan;
 pub(crate) mod logical_planner;
